CA: extract TestClusterStateRegistryScaleUpWithDeletedNodes logic to a helper
The logic in TestClusterStateRegistryScaleUpWithDeletedNodes is
extracted to a new RunTestClusterStateRegistryScaleUpWithDeletedNodes
helper function that is run from the test.

Subsequent commits will make RunTestClusterStateRegistryScaleUpWithDeletedNodes
parametrizable with different CloudProvider implementations. This is an
intermediate step for easier review - the test logic is copied as-is
to the new helper in this commit.
